feat : 修改usb通信

+++ src/nav_transporter/include/usb.hpp
@@ -97,10 +97,10 @@ typedef struct
 	float LeftMotorAngle;
 	// imu数据
 	uint32_t TimeStamp;
-  int16_t q0;
-  int16_t q1;
-  int16_t q2;
-  int16_t q3;
+  float q0;
+  float q1;
+  float q2;
+  float q3;
   // 包尾
   uint8_t _EOF;
 } NavIMUReceivePackage;
@@ -127,10 +127,10 @@ typedef struct
     uint8_t ID;
 		// imu数据
 		uint32_t TimeStamp;
-    int16_t q0;
-    int16_t q1;
-    int16_t q2;
-    int16_t q3;
+    float q0;
+    float q1;
+    float q2;
+    float q3;
     // 包尾
     uint8_t _EOF;
 } RMOSIMUReceivePackage;
